requirements.txt
biopython==1.65
Cython==0.22
jdcal==1.0
nose==1.3.4
numpy==1.9.2
openpyxl==2.2.0
pandas==0.15.2
primer3-py==0.4.1
python-dateutil==2.4.1
pytz==2014.10
six==1.9.0

-e git+git://github.com/Wyss/libnano#egg=libnano
-e git+git://github.com/Wyss/mauve-py#egg=mauve-py
